6.6.0

+++ src/corelib/io/qstorageinfo_p.h
@@ -40,7 +40,7 @@ protected:
     void retrieveDiskFreeSpace();
     bool queryStorageProperty();
     void queryFileFsSectorSizeInformation();
-#elif defined(Q_OS_MAC)
+#elif defined(Q_OS_DARWIN)
     void retrievePosixInfo();
     void retrieveUrlProperties(bool initRootPath = false);
     void retrieveLabel();
@@ -58,7 +58,7 @@ public:
     qint64 bytesTotal;
     qint64 bytesFree;
     qint64 bytesAvailable;
-    int blockSize;
+    ulong blockSize;
 
     bool readOnly;
     bool ready;
